temporary path velocity solution
The spline derivatives are not accurate due to uneven time steps. Added potential solutions in "future work" at the bottom of planning.py. Implemented a temporary solution so at least the integral of velocity gives the position. The velocities themselves are not calculated correctly because the spline assumes that every segment is 1s long. Therefore, if you have to go 100m in 100s, the spline will speed up a lot in the middle, when it should just stay at 1m/s the whole way.